+++ modeler/example_test.go
@@ -10,7 +10,7 @@ import (
 
 func Example() {
 	doc := gltf.NewDocument()
-	positionAccessor := modeler.WritePosition(doc, [][3]float32{{43, 43, 0}, {83, 43, 0}, {63, 63, 40}, {43, 83, 0}, {83, 83, 0}})
+	positionAccessor := modeler.WritePosition(doc, [][3]float64{{43, 43, 0}, {83, 43, 0}, {63, 63, 40}, {43, 83, 0}, {83, 83, 0}})
 	indicesAccessor := modeler.WriteIndices(doc, []uint16{0, 1, 2, 3, 1, 0, 0, 2, 3, 1, 4, 2, 4, 3, 2, 4, 1, 3})
 	colorIndices := modeler.WriteColor(doc, [][3]uint8{{50, 155, 255}, {0, 100, 200}, {255, 155, 50}, {155, 155, 155}, {25, 25, 25}})
 	doc.Meshes = []*gltf.Mesh{{
@@ -39,8 +39,8 @@ func ExampleWriteImage() {
 	}
 	doc := gltf.NewDocument()
 	indicesAccessor := modeler.WriteIndices(doc, []uint16{0, 1, 2, 3, 1, 0, 0, 2, 3, 1, 4, 2, 4, 3, 2, 4, 1, 3})
-	positionAccessor := modeler.WritePosition(doc, [][3]float32{{43, 43, 0}, {83, 43, 0}, {63, 63, 40}, {43, 83, 0}, {83, 83, 0}})
-	textureAccessor := modeler.WriteTextureCoord(doc, [][2]float32{{0, 1}, {0.4, 1}, {0.4, 0}, {0.4, 1}, {0, 1}})
+	positionAccessor := modeler.WritePosition(doc, [][3]float64{{43, 43, 0}, {83, 43, 0}, {63, 63, 40}, {43, 83, 0}, {83, 83, 0}})
+	textureAccessor := modeler.WriteTextureCoord(doc, [][2]float64{{0, 1}, {0.4, 1}, {0.4, 0}, {0.4, 1}, {0, 1}})
 	imageIdx, err := modeler.WriteImage(doc, "gopher", "image/png", bytes.NewReader(img))
 	if err != nil {
 		panic(err)
@@ -78,7 +78,7 @@ func ExampleWriteImage() {
 func ExampleWriteAccessorsInterleaved() {
 	doc := gltf.NewDocument()
 	indices, _ := modeler.WriteAccessorsInterleaved(doc,
-		[][3]float32{{43, 43, 0}, {83, 43, 0}, {63, 63, 40}, {43, 83, 0}, {83, 83, 0}},
+		[][3]float64{{43, 43, 0}, {83, 43, 0}, {63, 63, 40}, {43, 83, 0}, {83, 83, 0}},
 		[][3]uint8{{50, 155, 255}, {0, 100, 200}, {255, 155, 50}, {155, 155, 155}, {25, 25, 25}},
 	)
 	indicesAccessor := modeler.WriteIndices(doc, []uint16{0, 1, 2, 3, 1, 0, 0, 2, 3, 1, 4, 2, 4, 3, 2, 4, 1, 3})
@@ -104,11 +104,11 @@ func ExampleWriteAccessorsInterleaved() {
 func ExampleWriteAttributesInterleaved() {
 	doc := gltf.NewDocument()
 	attrs, _ := modeler.WriteAttributesInterleaved(doc, modeler.Attributes{
-		Position:       [][3]float32{{1, 2, 3}, {0, 0, -1}},
-		Normal:         [][3]float32{{1, 2, 3}, {0, 0, -1}},
-		Tangent:        [][4]float32{{1, 2, 3, 4}, {1, 2, 3, 4}},
+		Position:       [][3]float64{{1, 2, 3}, {0, 0, -1}},
+		Normal:         [][3]float64{{1, 2, 3}, {0, 0, -1}},
+		Tangent:        [][4]float64{{1, 2, 3, 4}, {1, 2, 3, 4}},
 		TextureCoord_0: [][2]uint8{{0, 255}, {255, 0}},
-		TextureCoord_1: [][2]float32{{1, 2}, {1, 2}},
+		TextureCoord_1: [][2]float64{{1, 2}, {1, 2}},
 		Joints:         [][4]uint8{{1, 2, 3, 4}, {1, 2, 3, 4}},
 		Weights:        [][4]uint8{{1, 2, 3, 4}, {1, 2, 3, 4}},
 		Color:          [][3]uint8{{255, 255, 255}, {0, 255, 0}},
